+++ UI.Web/ViewModels/Stock/StockArticuloSucursalViewModel.cs
@@ -9,7 +9,7 @@ namespace UI.Web.ViewModels.Stock
         public virtual Articulo Articulo { get; set; }
         public virtual Sucursal Sucursal { get; set; }
         public int SucursalID { get; set; }
-        public int StockActual { get; set; }
-        public int StockAgregar { get; set; }
+        public decimal StockActual { get; set; }
+        public decimal StockAgregar { get; set; }
     }
 }
