+++ UI.Web/ViewModels/Stock/StockIndexViewModel.cs
@@ -10,11 +10,11 @@ namespace UI.Web.ViewModels.Stock
         {
             Articulos = new List<Articulo>();
             Sucursales = new List<Sucursal>();
-            TotalesStock = new List<int>();
+            TotalesStock = new List<decimal>();
         }
 
         public List<Articulo> Articulos { get; set; }
-        public List<int> TotalesStock { get; set; }
+        public List<decimal> TotalesStock { get; set; }
         public List<Sucursal> Sucursales { get; set; }
         public bool MostrarSoloAlertas { get; set; }
